refactor: move OfferSortingForm to proper folder
